LayoutTests/http/tests/webxr/webxr-third-party-iframe-issessionsupported-allowed-by-feature-policy.html

<!DOCTYPE html>
<html>
<body>
<script src="/js-test-resources/js-test.js"></script>
<script>
description("Tests that third-party iframes will be allowed webxr session support with xr-spatial-tracking feature policy.");
jsTestIsAsync = true;

onmessage = (_e) => {
    e = _e;
    if (e.data.isSessionSupported != undefined) {
        shouldBeTrue("e.data.isSessionSupported");
        finishJSTest();
    }
};
</script>
<iframe src="http://localhost:8000/webxr/resources/webxr-issessionsupported-test.html" allow="xr-spatial-tracking"></iframe>
</body>
</html>
