LayoutTests/http/tests/webxr/webxr-third-party-iframe-issessionsupported-allowed-by-feature-policy-expected.txt

Tests that third-party iframes will be allowed webxr session support with xr-spatial-tracking feature policy.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS e.data.isSessionSupported is true
PASS successfullyParsed is true

TEST COMPLETE
